Got many to many relationship to work. Fixed issue with same leader being used as an option

+++ Component-Info-API/src/main/java/com/starwars/rebellion/ComponentInfoAPI/repositorys/ActionCardRepository.java
@@ -3,6 +3,6 @@ package com.starwars.rebellion.ComponentInfoAPI.repositorys;
 import com.starwars.rebellion.ComponentInfoAPI.dao.ActionCard;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ActionCardRepository extends CrudRepository<ActionCard, Long> {
+public interface ActionCardRepository extends CrudRepository<ActionCard, Integer> {
     ActionCard findByCardTextTitle(String title);
 }
